+++ src/main/java/com/sofodev/armorplus/config/ArmorPlusConfig.java
@@ -21,10 +21,10 @@ public class ArmorPlusConfig {
     public static MaterialConfig infusedLavaMaterial;
     public static MaterialConfig frostMaterial;
     public static MaterialConfig frostLavaMaterial;
-    public static MaterialConfig guardianMaterial;
-    public static MaterialConfig superStarMaterial;
-    public static MaterialConfig enderDragonMaterial;
-    public static MaterialConfig slayerMaterial;
+    public static AdvancedMaterialConfig guardianMaterial;
+    public static AdvancedMaterialConfig superStarMaterial;
+    public static AdvancedMaterialConfig enderDragonMaterial;
+    public static AdvancedMaterialConfig slayerMaterial;
     public static MaterialConfig enhancedMaterial;
     //Advancements
     public static ConfigValueListener<Boolean> enableThankYouAdvancement;
@@ -81,10 +81,22 @@ public class ArmorPlusConfig {
         builder.pop(2);
         frostMaterial = new MaterialConfig(builder, subscriber, "frost");
         frostLavaMaterial = new MaterialConfig(builder, subscriber, "frost_lava");
-        guardianMaterial = new MaterialConfig(builder, subscriber, "guardian");
-        superStarMaterial = new MaterialConfig(builder, subscriber, "super_star");
-        enderDragonMaterial = new MaterialConfig(builder, subscriber, "ender_dragon");
-        slayerMaterial = new MaterialConfig(builder, subscriber, "slayer");
+        guardianMaterial = new AdvancedMaterialConfig(builder, subscriber, "guardian", true, 0, true, 0, false, 0,
+                false, 0, false, 0, true, 1, false, 0, false, 0,
+                false, 0, false, 0, false, 0, false, 0, false, 0,
+                false, 0, false, false, false, true);
+        superStarMaterial = new AdvancedMaterialConfig(builder, subscriber, "super_star", false, 0, false, 0, false, 0,
+                false, 0, false, 0, false, 0, true, 0, false, 0,
+                true, 0, false, 0, false, 0, false, 0, true, 0,
+                false, 0, false, true, false, true);
+        enderDragonMaterial = new AdvancedMaterialConfig(builder, subscriber, "ender_dragon", false, 0, false, 0, false, 0,
+                false, 0, false, 0, false, 0, false, 0, false, 0,
+                true, 0, false, 0, false, 0, false, 0, false, 0,
+                true, 0, true, true, false, true);
+        slayerMaterial = new AdvancedMaterialConfig(builder, subscriber, "slayer", false, 0, true, 0, false, 0,
+                false, 0, false, 0, true, 1, true, 0, false, 0,
+                true, 0, false, 0, false, 0, false, 0, true, 0,
+                true, 0, true, true, false, true);
         enhancedMaterial = new MaterialConfig(builder, subscriber, "enhanced");
         builder.pop();
         builder.push("world_gen");
@@ -152,6 +164,120 @@ public class ArmorPlusConfig {
             builder.pop();
         }
     }
+
+
+    public static class AdvancedMaterialConfig extends MaterialConfig {
+        public ConfigValueListener<Boolean> enableNIGHT_VISION;         // Index 0
+        public ConfigValueListener<Integer> amplifierNIGHT_VISION;      // Index 0
+        public ConfigValueListener<Boolean> enableWATER_BREATHING;      // Index 1
+        public ConfigValueListener<Integer> amplifierWATER_BREATHING;   // Index 1
+        public ConfigValueListener<Boolean> enableSTRENGTH;             // Index 2
+        public ConfigValueListener<Integer> amplifierSTRENGTH;          // Index 2
+        public ConfigValueListener<Boolean> enableSPEED;                // Index 3
+        public ConfigValueListener<Integer> amplifierSPEED;             // Index 3
+        public ConfigValueListener<Boolean> enableHASTE;                // Index 4
+        public ConfigValueListener<Integer> amplifierHASTE;             // Index 4
+        public ConfigValueListener<Boolean> enableJUMP_BOOST;           // Index 5
+        public ConfigValueListener<Integer> amplifierJUMP_BOOST;        // Index 5
+        public ConfigValueListener<Boolean> enableREGENERATION;         // Index 6
+        public ConfigValueListener<Integer> amplifierREGENERATION;      // Index 6
+        public ConfigValueListener<Boolean> enableRESISTANCE;           // Index 7
+        public ConfigValueListener<Integer> amplifierRESISTANCE;        // Index 7
+        public ConfigValueListener<Boolean> enableFIRE_RESISTANCE;      // Index 8
+        public ConfigValueListener<Integer> amplifierFIRE_RESISTANCE;   // Index 8
+        public ConfigValueListener<Boolean> enableSATURATION;           // Index 9
+        public ConfigValueListener<Integer> amplifierSATURATION;        // Index 9
+        public ConfigValueListener<Boolean> enableINVISIBILITY;         // Index 10
+        public ConfigValueListener<Integer> amplifierINVISIBILITY;      // Index 10
+        public ConfigValueListener<Boolean> enableHEALTH_BOOST;         // Index 11
+        public ConfigValueListener<Integer> amplifierHEALTH_BOOST;      // Index 11
+        public ConfigValueListener<Boolean> enableABSORPTION;           // Index 12
+        public ConfigValueListener<Integer> amplifierABSORPTION;        // Index 12
+        public ConfigValueListener<Boolean> enableSLOW_FALLING;         // Index 13
+        public ConfigValueListener<Integer> amplifierSLOW_FALLING;      // Index 13
+        public ConfigValueListener<Boolean> enableFLIGHT;               // Index 14
+        public ConfigValueListener<Boolean> enableWITHER_IMMUNITY;      // Index 15
+        public ConfigValueListener<Boolean> enableNATURAL_IMMUNITY;     // Index 16
+        public ConfigValueListener<Boolean> enableFIRE_EXTINGUISH;      // Index 17
+
+        public AdvancedMaterialConfig(ForgeConfigSpec.Builder builder, ConfigHelper.Subscriber subscriber, String name,
+                boolean enableNightVision, int amplifierNightVision, boolean enableWaterBreathing, int amplifierWaterBreathing, boolean enableStrength, int amplifierStrength,
+                boolean enableSpeed, int amplifierSpeed, boolean enableHaste, int amplifierHaste, boolean enableJumpBoost, int amplifierJumpBoost,
+                boolean enableRegeneration, int amplifierRegeneration, boolean enableResistance, int amplifierResistance, boolean enableFireResistance, int amplifierFireResistance,
+                boolean enableSaturation, int amplifierSaturation, boolean enableInvisibility, int amplifierInvisibility, boolean enableHealthBoost, int amplifierHealthBoost,
+                boolean enableAbsorption, int amplifierAbsorption, boolean enableSlowFalling, int amplifierSlowFalling,
+                boolean enableFlight, boolean enableWitherImmunity, boolean enableNaturalImmunity, boolean enableFireExtinguish) {
+            super(builder, subscriber, name);
+            builder.comment(name + " extended equipment Configuration")
+                    .push(name).push("effects");
+            enableNIGHT_VISION = subscriber.subscribe(builder.comment(name + " armor: enable/disable night vision")
+                    .define("enableNIGHT_VISION", enableNightVision));
+            amplifierNIGHT_VISION = subscriber.subscribe(builder.comment(name + " armor: night vision amplifier")
+                    .define("amplifierNIGHT_VISION", amplifierNightVision));
+            enableWATER_BREATHING = subscriber.subscribe(builder.comment(name + " armor: enable/disable water breathing")
+                    .define("enableWATER_BREATHING", enableWaterBreathing));
+            amplifierWATER_BREATHING = subscriber.subscribe(builder.comment(name + " armor: water breathing amplifier")
+                    .define("amplifierWATER_BREATHING", amplifierWaterBreathing));
+            enableSTRENGTH = subscriber.subscribe(builder.comment(name + " armor: enable/disable strength")
+                    .define("enableSTRENGTH", enableStrength));
+            amplifierSTRENGTH = subscriber.subscribe(builder.comment(name + " armor: strength amplifier")
+                    .define("amplifierSTRENGTH", amplifierStrength));
+            enableSPEED = subscriber.subscribe(builder.comment(name + " armor: enable/disable speed")
+                    .define("enableSPEED", enableSpeed));
+            amplifierSPEED = subscriber.subscribe(builder.comment(name + " armor: speed amplifier")
+                    .define("amplifierSPEED", amplifierSpeed));
+            enableHASTE = subscriber.subscribe(builder.comment(name + " armor: enable/disable haste")
+                    .define("enableHASTE", enableHaste));
+            amplifierHASTE = subscriber.subscribe(builder.comment(name + " armor: haste amplifier")
+                    .define("amplifierHASTE", amplifierHaste));
+            enableJUMP_BOOST = subscriber.subscribe(builder.comment(name + " armor: enable/disable jump boost")
+                    .define("enableJUMP_BOOST", enableJumpBoost));
+            amplifierJUMP_BOOST = subscriber.subscribe(builder.comment(name + " armor: jump boost amplifier")
+                    .define("amplifierJUMP_BOOST", amplifierJumpBoost));
+            enableREGENERATION = subscriber.subscribe(builder.comment(name + " armor: enable/disable regeneration")
+                    .define("enableREGENERATION", enableRegeneration));
+            amplifierREGENERATION = subscriber.subscribe(builder.comment(name + " armor: regeneration amplifier")
+                    .define("amplifierREGENERATION", amplifierRegeneration));
+            enableRESISTANCE = subscriber.subscribe(builder.comment(name + " armor: enable/disable resistance")
+                    .define("enableRESISTANCE", enableResistance));
+            amplifierRESISTANCE = subscriber.subscribe(builder.comment(name + " armor: resistance amplifier")
+                    .define("amplifierRESISTANCE", amplifierResistance));
+            enableFIRE_RESISTANCE = subscriber.subscribe(builder.comment(name + " armor: enable/disable fire resistance")
+                    .define("enableFIRE_RESISTANCE", enableFireResistance));
+            amplifierFIRE_RESISTANCE = subscriber.subscribe(builder.comment(name + " armor: fire resistance amplifier")
+                    .define("amplifierFIRE_RESISTANCE", amplifierFireResistance));
+            enableSATURATION = subscriber.subscribe(builder.comment(name + " armor: enable/disable saturation")
+                    .define("enableSATURATION", enableSaturation));
+            amplifierSATURATION = subscriber.subscribe(builder.comment(name + " armor: saturation amplifier")
+                    .define("amplifierSATURATION", amplifierSaturation));
+            enableINVISIBILITY = subscriber.subscribe(builder.comment(name + " armor: enable/disable invisibility")
+                    .define("enableINVISIBILITY", enableInvisibility));
+            amplifierINVISIBILITY = subscriber.subscribe(builder.comment(name + " armor: invisibility amplifier")
+                    .define("amplifierINVISIBILITY", amplifierInvisibility));
+            enableHEALTH_BOOST = subscriber.subscribe(builder.comment(name + " armor: enable/disable health boost")
+                    .define("enableHEALTH_BOOST", enableHealthBoost));
+            amplifierHEALTH_BOOST = subscriber.subscribe(builder.comment(name + " armor: health boost amplifier")
+                    .define("amplifierHEALTH_BOOST", amplifierHealthBoost));
+            enableABSORPTION = subscriber.subscribe(builder.comment(name + " armor: enable/disable absorption")
+                    .define("enableABSORPTION", enableAbsorption));
+            amplifierABSORPTION = subscriber.subscribe(builder.comment(name + " armor: absorption amplifier")
+                    .define("amplifierABSORPTION", amplifierAbsorption));
+            enableSLOW_FALLING = subscriber.subscribe(builder.comment(name + " armor: enable/disable slow falling")
+                    .define("enableSLOW_FALLING", enableSlowFalling));
+            amplifierSLOW_FALLING = subscriber.subscribe(builder.comment(name + " armor: slow falling amplifier")
+                    .define("amplifierSLOW_FALLING", amplifierSlowFalling));
+            enableFLIGHT = subscriber.subscribe(builder.comment(name + " armor: enable/disable flight")
+                    .define("enableFLIGHT", enableFlight));
+            enableWITHER_IMMUNITY = subscriber.subscribe(builder.comment(name + " armor: enable/disable wither immunity")
+                    .define("enableWITHER_IMMUNITY", enableWitherImmunity));
+            enableNATURAL_IMMUNITY = subscriber.subscribe(builder.comment(name + " armor: enable/disable natural immunity")
+                    .define("enableNATURAL_IMMUNITY", enableNaturalImmunity));
+            enableFIRE_EXTINGUISH = subscriber.subscribe(builder.comment(name + " armor: enable/disable fire extinguish")
+                    .define("enableFIRE_EXTINGUISH", enableFireExtinguish));
+            builder.pop().pop();
+        }
+    }
+
 
     public static class DropConfig {
         public ConfigValueListener<Boolean> enableRegularDrops;